refactor: move configs to a common subfolder
