+++ configs/cheminfo.js
