+++ eslint.config.js
@@ -1,5 +1,5 @@
 import { defineConfig } from 'eslint/config';
 
-import cheminfo from './cheminfo.js';
+import cheminfo from './configs/cheminfo.js';
 
 export default defineConfig(cheminfo);
