+++ package.json
@@ -4,16 +4,13 @@
   "description": "Shared ESLint config for cheminfo and ml.js projects",
   "type": "module",
   "exports": {
-    ".": "./cheminfo.js",
-    "./base": "./base.js",
-    "./jsdoc": "./jsdoc.js",
-    "./unicorn": "./unicorn.js"
+    ".": "./configs/cheminfo.js",
+    "./base": "./configs/base.js",
+    "./jsdoc": "./configs/jsdoc.js",
+    "./unicorn": "./configs/unicorn.js"
   },
   "files": [
-    "cheminfo.js",
-    "base.js",
-    "jsdoc.js",
-    "unicorn.js"
+    "configs"
   ],
   "scripts": {
     "eslint": "eslint *.js test/test.js",
@@ -35,6 +32,7 @@
   },
   "homepage": "https://github.com/cheminfo/eslint-config#readme",
   "dependencies": {
+    "@vitest/eslint-plugin": "^1.3.4",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-jsdoc": "^51.3.3",
     "eslint-plugin-unicorn": "^59.0.1",
